+++ libs/causality/src/lib/CausalAnalysisDashboard/Controls/CausalAnalysisView/CausalIndividualView/CausalIndividualChart.tsx
@@ -158,16 +158,14 @@ export class CausalIndividualChart extends React.PureComponent<
               </MissingParametersPlaceholder>
             )}
             {canRenderChart && (
-              <div id="highchartContainer">
-                <BasicHighChart
-                  configOverride={getIndividualChartOptions(
-                    plotlyProps,
-                    this.selectPointFromChart
-                  )}
-                  theme={getTheme()}
-                  id="CausalAggregateChart"
-                />
-              </div>
+              <BasicHighChart
+                configOverride={getIndividualChartOptions(
+                  plotlyProps,
+                  this.selectPointFromChart
+                )}
+                theme={getTheme()}
+                id="CausalAggregateChart"
+              />
             )}
           </div>
           <div className={classNames.horizontalAxisWithPadding}>
@@ -379,7 +377,7 @@ export class CausalIndividualChart extends React.PureComponent<
       const metaX =
         this.context.jointDataset.metaDict[chartProps.xAxis.property];
       const rawX = JointDataset.unwrap(dictionary, chartProps.xAxis.property);
-      hovertemplate += `${metaX.label}: %{customdata.X}<br>`;
+      hovertemplate += `${metaX.label}: {point.customdata.X}<br>`;
 
       rawX.forEach((val, index) => {
         if (metaX.treatAsCategorical) {
@@ -406,7 +404,7 @@ export class CausalIndividualChart extends React.PureComponent<
       const metaY =
         this.context.jointDataset.metaDict[chartProps.yAxis.property];
       const rawY = JointDataset.unwrap(dictionary, chartProps.yAxis.property);
-      hovertemplate += `${metaY.label}: %{customdata.Y}<br>`;
+      hovertemplate += `${metaY.label}: {point.customdata.Y}<br>`;
       rawY.forEach((val, index) => {
         if (metaY.treatAsCategorical) {
           customdata[index].Y = metaY.sortedCategoricalValues?.[val];
@@ -428,7 +426,7 @@ export class CausalIndividualChart extends React.PureComponent<
         trace.y = rawY;
       }
     }
-    hovertemplate += `${localization.Interpret.Charts.rowIndex}: %{customdata.Index}<br>`;
+    hovertemplate += `${localization.Interpret.Charts.rowIndex}: {point.customdata.Index}<br>`;
     hovertemplate += "<extra></extra>";
     trace.customdata = customdata as any;
     trace.hovertemplate = hovertemplate;
