+++ libs/core-ui/src/index.ts
@@ -75,3 +75,4 @@ export * from "./lib/Interfaces/IErrorAnalysisData";
 export * from "./lib/Highchart/BasicHighChart";
 export * from "./lib/Highchart/FeatureImportanceDependence";
 export * from "./lib/Highchart/FeatureImportanceBar";
+export * from "./lib/Highchart/IHighchartsConfig";
