+++ libs/core-ui/src/lib/Highchart/ICommonChartProps.ts
@@ -3,7 +3,7 @@
 
 import { ITheme } from "@fluentui/react";
 
-import { IHighchartsConfig } from "./HighchartTypes";
+import { IHighchartsConfig } from "./IHighchartsConfig";
 
 export interface ICommonChartProps {
   id?: string;
