+++ libs/core-ui/src/lib/util/getDependencyChartOptions.ts
@@ -3,7 +3,7 @@
 
 import { ITheme } from "@fluentui/react";
 
-import { IHighchartsConfig } from "../Highchart/HighchartTypes";
+import { IHighchartsConfig } from "../Highchart/IHighchartsConfig";
 
 export interface IDependenceData {
   x: number;
