+++ libs/core-ui/src/lib/util/getFeatureImportanceBoxOptions.ts
@@ -4,7 +4,7 @@
 import { ITheme } from "@fluentui/react";
 
 import { IGlobalSeries } from "../Highchart/FeatureImportanceBar";
-import { IHighchartsConfig } from "../Highchart/HighchartTypes";
+import { IHighchartsConfig } from "../Highchart/IHighchartsConfig";
 
 import { FabricStyles } from "./FabricStyles";
 import { getBoxData } from "./getBoxData";
